+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.0.4] - 2023-10-10
+
 ### Changed
 
 - Refactor alert component factory methods
@@ -83,7 +85,8 @@ Stay tuned for more to come...
 
 - for dependency upgrades
 -->
-[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/patternfly-java/patternfly-java/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/patternfly-java/patternfly-java/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/patternfly-java/patternfly-java/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/patternfly-java/patternfly-java/compare/vTemplate...v0.0.1